+++ R/proxy.R
@@ -2,12 +2,10 @@
 #'
 #' Tries to find the local proxy settings to access a given url.
 #'
-#' @param url The url to access, as a character string.
-#'
-#' @return A list of proxy, or \code{NULL} if not proxy are found.
+#' @return A proxy, or \code{NULL} if not proxy are found.
 #' @keywords internal
-get_proxy_list <- function(url) {
-  proxy_string <- curl::ie_get_proxy_for_url(target_url = url)
+get_proxy <- function() {
+  proxy_string <- options()$rgho.http_proxy
 
   if (is.null(proxy_string)) {
     return(NULL)
@@ -26,13 +24,24 @@ get_proxy_list <- function(url) {
 #' @return A list of proxy settings.
 #' @keywords internal
 parse_proxy_string <- function(proxy_string) {
-  proxy_string %>%
-    strsplit(";") %>%
-    magrittr::extract2(1) %>%
-    strsplit(":") %>%
-    lapply(
-      function(x) {
-        httr::use_proxy(x[1], as.integer(x[2]))
-      }
-    )
+  list_proxy <- httr::parse_url(proxy_string)
+
+  if (is.null(list_proxy$hostname)) {
+    stop("Invalid proxy URL.")
+  }
+
+  list_proxy$url <- list_proxy$hostname
+
+  list_proxy$port <- as.integer(list_proxy$port)
+
+  if (is.null(list_proxy$port)) {
+    stop("Invalid proxy port.")
+  }
+
+  do.call(
+    httr::use_proxy,
+    list_proxy[intersect(
+      names(list_proxy),
+      c("url", "port", "username", "password"))]
+  )
 }